+++ apps/admin/app/(main)/event-manage/_components/step-event/poster-image-field.tsx
