+++ apps/admin/app/(main)/event-manage/_components/step-event/thumbnail-image-field.tsx
